requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0

Babel>=1.3
cliff-tablib>=1.0
cliff>=1.7.0  # Apache-2.0
os-cloud-config
python-heatclient>=0.3.0
python-ironicclient>=0.4.1
python-openstackclient>=1.0.0
tripleo-common

# The ironic-discoverd OSC integration isn't yet on PyPI
-e git://github.com/stackforge/ironic-discoverd.git#egg=ironic_discoverd
